Add void return type to test functions
Added return type declarations to test functions for better code reliability and understanding the code flow. This ensures each with PHPUnit test function does not return a value as recommended by PHPUnit's best practice. The primary reason for this change is to enforce the use of assertions over returned values.